booking optimization

+++ app/src/main/java/com/footzone/footzone/model/StadiumDataToBottomSheetDialog.kt
@@ -2,6 +2,6 @@ package com.footzone.footzone.model
 
 data class StadiumDataToBottomSheetDialog(
     val stadiumId: String,
-    val hourlyPrice: Long,
+    val hourlyPrice: Int,
     val workingDays: List<WorkingDay>
 )
